@@ -2,11 +2,11 @@
 
 ext {
 
-    android = [compileSdkVersion   : 28,
+    android = [compileSdkVersion   : 29,
                buildToolsVersion   : "28.0.3",
                applicationId       : "com.seniorlibs.app",
                minSdkVersion       : 15,
-               targetSdkVersion    : 28,
+               targetSdkVersion    : 29,
                versionCode         : 1,
                versionName         : "1.0.0",
                multiDexEnabled     : true,
