@@ -1,7 +1,7 @@
 package com.seniorlibs.app.grammar;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.seniorlibs.app.R;
 
